+++ Runtime/Scripts/Components/UIs/UIRoot.cs
@@ -1,0 +1,4 @@
+﻿namespace ParkMinPackages.Workflow.Default.Components.UIs
+{
+	public class UIRoot : Actor { }
+}